fixed a styling issue.
